rework dfn -> toml conversion
TOML conversion as per group discussions and as prototyped in flopy3 in the Jinja PR.

This is a nested structure where components have blocks, which have variables, which may be composites of other variables.

Add the latest DFNs as well as the converted TOML files.

The structure is still tentative and may evolve.