minimal lark parser (#175)
generic parser just recognizing blocks and lines within them. while a type-aware parser is coming soon, we can fall back to this in case type info is ever not available. if calling load on simulation, model or package, it will be. but maybe we also want a generic load that tries to infer the component from the input file name and contents, in which case we can connect this to a type-aware converter that can match fields against known types to recognize the component, then parse fields before structuring?

also eagerly import submodules of flopy.mf6 so components are all registered on first import of the top-level module.