add to_dict method to component (#212)
supports either fields or blocks (nested). useful for the writer internally and also presumably for users.

includes a hack to omit "derived" dimensions from the dict like nodes, which is referenced in DFNs (therefore by components in their array definitions) but isn't actually a variable in any DFN. a better way would be an option on the dim field decorator, maybe. come back to it.