factor out function for structured grid dims workaround (#215)
in lieu of an xarray index for structured grids to provide layer/row/column indexing on arrays even when they are defined in terms of nodes, we reshape flat arrays to structured grid dims in the converter for now. move this out into a function so it's more obvious it's temporary.

Co-authored-by: mjreno <renomik@gmail.com>