draft object dtype perioddata for oc (#223)
prototype a strongly-typed, more true-to-DFNs alternative to the imod-python style OC we've had so far. writing doesn't work yet. this could be generated from an OC DFN modified to look something like this. cf #205

also refactor the converter file into a module and rename dict_to_array to structure_array.