adapt time utilities from imod-python (#244)
Begin to implement the time side of #226 and also to address #243 with a Time subclass of flopy3's ModelTime with extra capabilities.

Adapt some time-related utilities from imod-python supporting the ability to generate a TDIS from timestamps

- https://github.com/Deltares/imod-python/blob/87bbed86a6e36b1b65644625aabc6c3d5b6fda63/imod/util/time.py#L98
- https://github.com/Deltares/imod-python/blob/87bbed86a6e36b1b65644625aabc6c3d5b6fda63/imod/mf6/simulation.py#L177

This can be the foundation for a similar method to create a TDIS from simulation packages later on.